==> scripts/STEP2_FDD_FooDB_Content_Mapping.Rmd <==
---
title: "Map Disaggregated Foods to FooDB"
author: "Stephanie Wilson"
date: "February 2026"
output:
  md_document:
    variant: gfm
    toc: true
    number_sections: false
    df_print: kable
  html_document:
    toc: true
    toc_depth: 3
    number_sections: false
    theme: sandstone
    highlight: tango
    df_print: paged
---

## Map Disaggregated Foods to FooDB
This script takes your disaggregated foods (from ASA24 or NHANES) and maps them to FooDB to derive polyphenol content. 

#### INPUTS

-   **Diet_Disaggregated.csv.bz2**: Input dietary data that has been disaggregated using FDD.
-   **FDA_FooDB_Mapping_Nov_2025.csv**: FDD to FooDB matches.\
-   **FooDB_polyphenol_content_with_dbPUPsubstrates_Aug25.csv.bz2**: Phenols pulled out of Compounds.csv and matched to FooDB's Compounds file with cleaned text descriptions. Includes dbPUP substrates
-   **FooDB_phenol_content_foodsums_Dec24Update.csv**: Summed polyphenol intake per unique food id in FooDB. Specific foods not present in FooDB or present but not quantified have had their concentrations adjusted.

#### OUTPUTS
- **Diet_Disaggregated_mapped.csv.bz2**; Disaggregated dietary data, mapped to FooDB foods
- **Diet_FooDB_polyphenol_content.csv.bz2**: Disaggregated dietary data, mapped to FooDB foods and polyphenol content
- **summary_missing_foods_overview.txt**: Summary of the number of unmapped foods between FDA-FDD and FooDB across ALL recalls
- **summary_missing_foods_detailed.csv**: Summary of the number of unmapped foods between FDA-FDD and FooDB BY recall

## SCRIPTS

```{r echo = FALSE}
# Clears all objects from memory
rm(list = ls())
```

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "..") # project root relative to scripts/
```

```{r warning = FALSE}
# Load packages
suppressMessages(library(dplyr))
suppressMessages(library(vroom))
suppressMessages(library(tidyr))
suppressMessages(library(stringr))
```

Load data
```{r}
# Load provided file paths
source("provided_files.R")

input = vroom::vroom('outputs/Diet_Disaggregated.csv.bz2', 
                     show_col_types = FALSE) %>%
  dplyr::select(-wweia_food_description)

# FDD to FooDB food mappings
mapping = vroom::vroom(mapping, show_col_types = FALSE) %>%
  dplyr::select(-c(method, score)) 

#FooDB polyphenol quantities
FooDB_mg_100g = vroom::vroom(FooDB_mg_100g, 
                     show_col_types = FALSE) %>%
  # Since we created orig_content_avg from multiple sources, ensure distinct values
  dplyr::distinct(food_id, compound_public_id, .keep_all = TRUE) %>%
  dplyr::select(-c(food_public_id, food_name)) %>%
  dplyr::relocate(orig_content_avg, .before = citation) %>%
  # Keep only quantified compounds
  dplyr::filter(!is.na(orig_content_avg_RFadj)) 
```

### Specify grouping variables
Column grouping depends on whether diet output is from a record or recall.
```{r}
if ("RecallNo" %in% names(input)) {
  group_vars = c("subject", "RecallNo")
  
} else if ("RecordNo" %in% names(input)) {
  group_vars = c("subject", "RecordNo", "RecordDayNo")
  
} else {
  stop("Data must contain RecallNo or RecordNo.")
}
```

### Connect Disaggregated ASA to FooDB through key link.

```{r}
input_mapped = input %>%
  # Connect to foodb names
  dplyr::left_join(mapping, by = c("fdd_ingredient"))

vroom::vroom_write(input_mapped, 'outputs/Diet_Disaggregated_mapped.csv.bz2', delim = ",")
```


### Merge FooDB-matched Ingredient Codes to FooDB Polyphenol Content File.

- Link between FooDB Polyphenol Content and code-matched data is *food_id*.
- Add *pp_consumed*, for polyphenol content (mg/100g multiply by 0.01 to get mg/g) by ingredient consumed (grams) to get the polyphenol amount consumed (mg).

```{r}
input_mapped_content = input_mapped %>%
  # Bring in the Polyphenol Content
  dplyr::left_join(FooDB_mg_100g, by = 'food_id', relationship = "many-to-many") %>%
  dplyr::select(-c(food_V2_ID.y, aggregate_RF)) %>%
  dplyr::rename(food_V2_ID = food_V2_ID.x) %>%
  # Calculate polyphenol amount consumed in milligrams
  # Specific Polyphenols in Tea from Duke and DFC seem to correspond to dry weight 
  # apply the correction for dry weight
  dplyr::mutate(
    pp_consumed = if_else(
      compound_public_id %in% c("FDB000095", "FDB017114") & food_id == 38,
      (orig_content_avg_RFadj * 0.01) * FoodAmt_Ing_g * (ingredient_percent / 100),
      (orig_content_avg_RFadj * 0.01) * FoodAmt_Ing_g))
```

Export polyphenol content file. Compress as this is the largest file that we generate.
```{r}
vroom::vroom_write(input_mapped_content, 'outputs/Diet_FooDB_polyphenol_content.csv.bz2', delim = ",")
```

### Review of Unmapped Foods

Find foods and food components that did not map to FooDB: 
```{r}
unmapped_foods = input_mapped %>% 
  dplyr::select(c(fdd_ingredient, orig_food_common_name)) %>% 
  # This extracts the entries that did not map
  dplyr::filter(!is.na(fdd_ingredient) & is.na(orig_food_common_name)) %>% 
  dplyr::distinct(fdd_ingredient, .keep_all = TRUE) %>%
  dplyr::pull(fdd_ingredient)

# Calculate summary statistics
percent_unmapped = (length(unmapped_foods)/length(mapping$fdd_ingredient))*100
```

How many recalls/records had at least one food missing
```{r}
#Count missing mappings for each recall
missing_counts = input_mapped %>%
  dplyr::group_by(across(all_of(group_vars))) %>%
  dplyr::summarise(
    missing = sum(is.na(orig_food_common_name)),
    total = n(),
    percent_missing = missing / total * 100,
    .groups = "drop")

# Calculate summary statistics
num_recalls_all_mapped = sum(missing_counts$percent_missing == 0)
num_recalls_some_missing = sum(missing_counts$percent_missing != 0)
```

Create a summary report reflecting the number of missing foods
```{r}
# Main Report on Missing Mappings
report_lines = c(
  "=== Unmapped Ingredients Report ===",
  paste("Report Time:", Sys.time()),
  paste("Number of unique FDA-FDD ingredients that did not map to FooDB:", length(unmapped_foods)),
  paste("Percentage of ingredients missing:", round(percent_unmapped, 2), "%"),
  "",
  "List of unmapped ingredients:",
  unmapped_foods,
  "",
  "=== Recall-level Missing Counts ===",
  paste("Number of records/recalls where all foods mapped:", num_recalls_all_mapped),
  paste("Number of records/recalls where at least one food did not map:", num_recalls_some_missing))

# Write to text file
writeLines(report_lines, "outputs/summary_missing_foods_overview.txt")

# Write out detailed recall-level data
vroom::vroom_write(missing_counts, 'outputs/summary_missing_foods_detailed.csv', delim = ",")
```
