==> scripts/STEP3d_Polyphenol_Summary_Food_Contributors.Rmd <==
---
title: "Food Contributors to Polyphenol Intake"
author: "Stephanie Wilson"
date: "February 2026"
output:
  md_document:
    variant: gfm
    toc: true
    number_sections: false
    df_print: kable
  html_document:
    toc: true
    toc_depth: 3
    number_sections: false
    theme: sandstone
    highlight: tango
    df_print: paged
---

## Food Contributors to Total Polyphenol Intake
This script examines food contributors to total polyphenol intake.

#### INPUTS
- **Diet_Disaggregated_mapped.csv.bz2**; Dissagregated dietary data, mapped to FooDB foods
- **Diet_FooDB_polyphenol_content.csv.bz2**: Disaggregated dietary data, mapped to FooDB foods and polyphenol content
- **Diet_total_nutrients.csv** - total daily nutrient data to go with dietary data.

#### OUTPUTS
- **summary_total_polyphenol_food_contributors.csv**

## SCRIPTS

```{r echo = FALSE, purl = FALSE}
# Clears all objects from memory
rm(list = ls())
```

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "..") # project root relative to scripts/
```

```{r warning = FALSE}
# Load packages
suppressMessages(library(dplyr))
suppressMessages(library(vroom))
suppressMessages(library(tidyr))
suppressMessages(library(stringr))
```

```{r}
# Load provided file paths
source("provided_files.R")

# Foods Mapped
input_mappings = vroom::vroom('outputs/Diet_Disaggregated_mapped.csv.bz2',
                              show_col_types = FALSE)

# Foods Mapped with content
input_polyphenol_content = vroom::vroom('outputs/Diet_FooDB_polyphenol_content.csv.bz2',
                                        show_col_types = FALSE)
input_kcal = vroom::vroom('outputs/Diet_total_nutrients.csv', show_col_types = FALSE) %>%
  # Ensure consistent KCAL naming whether ASA24 or NHANES
  dplyr::rename_with(~ "Total_KCAL", .cols = any_of(c("Total_KCAL", # Specific to ASA24
                                               "Total_DRXIKCAL"))) %>%  # Specific to NHANES
  dplyr::select(c(subject,
          # Ensures we pull correct columns for record or recall
           any_of(c("RecallNo", "RecordNo", "RecordDayNo")),
           Total_KCAL))

# Merge the two files
input_polyphenol_kcal = dplyr::left_join(input_polyphenol_content, input_kcal)
```

### Specify grouping variables
Column grouping depends on whether output is from a record or recall.
```{r}
if ("RecallNo" %in% names(input_polyphenol_kcal)) {
  group_vars = c("subject", "RecallNo", "fdd_ingredient")
  
} else if ("RecordNo" %in% names(input_polyphenol_kcal)) {
  group_vars = c("subject", "RecordNo", "RecordDayNo", "fdd_ingredient")
  
} else {
  stop("Data must contain RecallNo or RecordNo.")
}
```

### Food Consumption Counts
This tells us how many times a food was reported and then if we see all participants consume a food.
```{r}
food_counts = input_mappings %>%
  # Every time a subject consumed food, how much of it in total did they eat? 
  # How many instances did they eat this food in one day?
  dplyr::group_by(across(all_of(group_vars))) %>%
  dplyr::summarise(
    amount_consumed_g = sum(FoodAmt_Ing_g, na.rm =TRUE),
    times_consumed = n(),
    .groups = "drop") %>%
  
  # Across the cohort summary
  dplyr::group_by(fdd_ingredient) %>%
  dplyr::summarise(
    # How many instances was this food seen
    total_times_consumed = sum(times_consumed),
    
    # How many subjects consumed this food
    n_subjects = n_distinct(subject),
    
    # Information about the Ingredient consumed
    min_amount_consumed_g = min(amount_consumed_g),
    average_amount_consumed_g = mean(amount_consumed_g),
    median_amount_consumed_g= median(amount_consumed_g),
    max_amount_consumed_g = max(amount_consumed_g),
    
    .groups = "drop") %>%
  dplyr::arrange(desc(total_times_consumed))
```

### Total daily Polyphenol Intake Numbers BY RECALL

```{r}
content_by_entry_food = input_polyphenol_kcal %>%
  
  # Recall - Sum by Subject, Recall
  # Record - Sum by Subject, Record Number, Day in Record Number
  # Both recall and record group by compound
  dplyr::group_by(across(all_of(group_vars))) %>%
  dplyr::mutate(pp_recallsum_mg = sum(pp_consumed, na.rm = TRUE),
         pp_recallsum_mg1000kcal = pp_recallsum_mg/(Total_KCAL/1000)) %>%
  dplyr::ungroup() %>%
  dplyr::distinct(across(all_of(group_vars)), .keep_all = TRUE) %>%
  dplyr::select(c(subject, any_of(c("RecallNo", "RecordNo", "RecordDayNo")),
                  fdd_ingredient, pp_recallsum_mg, Total_KCAL, pp_recallsum_mg1000kcal))
```

### Total daily Polyphenol Intake Numbers AVERAGE FOR SUBJECT
```{r}
content_by_subject_food = content_by_entry_food %>%
  
  # Average by Participant
  dplyr::group_by(subject, fdd_ingredient) %>%
  dplyr::mutate(pp_average_mg = mean(pp_recallsum_mg, na.rm = TRUE),
         kcal_average = mean(Total_KCAL, na.rm = TRUE),
         pp_average_mg_1000kcal = pp_average_mg/(kcal_average/1000)) %>%
  dplyr::ungroup() %>%
  dplyr::distinct(subject, fdd_ingredient, .keep_all = TRUE) %>%
  dplyr::select(c(subject, fdd_ingredient, pp_average_mg, kcal_average, pp_average_mg_1000kcal)) 
```

### Obtain food contributors to total polyphenol intake
```{r}
content_by_food = content_by_subject_food %>%
  
  # Average by food
  dplyr::group_by(fdd_ingredient) %>%
  dplyr::mutate(food_pp_average_mg1000kcal = mean(pp_average_mg_1000kcal, na.rm = TRUE)) %>%
  dplyr::ungroup() %>%
  dplyr::distinct(fdd_ingredient, .keep_all = TRUE) %>%
  dplyr::select(c(fdd_ingredient, food_pp_average_mg1000kcal)) %>%
  dplyr::left_join(food_counts, by = "fdd_ingredient") %>%
  dplyr::arrange(desc(food_pp_average_mg1000kcal), desc(n_subjects))

# Export food contributors file 
vroom::vroom_write(content_by_food, 'outputs/summary_total_polyphenol_food_contributors.csv', delim = ",")
```
